requirements.txt
openai>=1.50.0
mysql-connector-python>=9.0.0
python-dotenv>=1.0.1
beautifulsoup4>=4.12.0
requests>=2.31.0